+++ app/agent/nodes/__init__.py
@@ -4,7 +4,6 @@ from app.agent.nodes.build_context import node_build_context
 from app.agent.nodes.extract_alert import node_extract_alert
 from app.agent.nodes.frame_problem.frame_problem import node_frame_problem
 from app.agent.nodes.investigate.node import node_investigate
-from app.agent.nodes.investigate.plan_actions.node import node_plan_actions
 from app.agent.nodes.publish_findings import node_publish_findings
 from app.agent.nodes.root_cause_diagnosis import node_diagnose_root_cause
 
@@ -13,7 +12,6 @@ __all__ = [
     "node_diagnose_root_cause",
     "node_extract_alert",
     "node_frame_problem",
-    "node_plan_actions",
     "node_investigate",
     "node_publish_findings",
 ]
