+++ app/agent/nodes/investigate/__init__.py
@@ -1,9 +1,7 @@
 """Investigate node package."""
 
 from app.agent.nodes.investigate.node import node_investigate
-from app.agent.nodes.investigate.plan_actions.node import node_plan_actions
 
 __all__ = [
-    "node_plan_actions",
     "node_investigate",
 ]
