+++ app/agent/nodes/investigate/plan_actions/__init__.py
@@ -1,6 +1,5 @@
 """Plan actions module."""
 
-from app.agent.nodes.investigate.plan_actions.node import node_plan_actions
 from app.agent.nodes.investigate.plan_actions.plan_actions import plan_actions
 
-__all__ = ["node_plan_actions", "plan_actions"]
+__all__ = ["plan_actions"]
